requirements.txt
python-dateutil ~= 2.4.2
pyyaml ~= 3.11.0
relatorio ~= 0.6.2
